split evdev and pynput in separate files
